move path parameters to project common
