+++ api/dependencies/parameters/identifiers.py
@@ -1,6 +1,12 @@
-from fastapi import Query
+from fastapi import Path, Query
 
-async def variant_identifier_param(variant: str = Query(regex="", description="")):
+from api.common.formatters import clean
+
+async def path_variant_id(variant: str = Query(regex="", description="")):
     return True
 
+async def path_track_id(track: str = Path(description="data track identifier")) -> str:
+    return clean(track)
 
+async def path_collection_name(collection: str = Path(description="FILER collection name")) -> str:
+    return clean(collection)
