--- api/dependencies/parameters/path.py
@@ -1,7 +1,0 @@
-
-from fastapi import Path
-
-from api.common.formatters import clean
-
-async def path_track_id(track: str = Path(description="data track identifier")) -> str:
-    return clean(track)
